moved optimisers implementation to src
