+++ libs/ml/include/ml/optimisation/momentum_optimiser.hpp
@@ -68,89 +68,6 @@ private:
   void Init();
 };
 
-template <class T>
-void MomentumOptimiser<T>::Init()
-{
-  for (auto &train : this->graph_trainables_)
-  {
-    this->momentum_.emplace_back(TensorType(train->GetWeights().shape()));
-  }
-  ResetMomentum();
-}
-
-template <class T>
-MomentumOptimiser<T>::MomentumOptimiser(std::shared_ptr<Graph<T>>       graph,
-                                        std::vector<std::string> const &input_node_names,
-                                        std::string const &             label_node_name,
-                                        std::string const &             output_node_name,
-                                        DataType const &                learning_rate,
-                                        DataType const &                momentum_update)
-  : Optimiser<T>(graph, input_node_names, label_node_name, output_node_name, learning_rate)
-  , momentum_update_(momentum_update)
-{
-  Init();
-}
-
-template <class T>
-MomentumOptimiser<T>::MomentumOptimiser(
-    std::shared_ptr<Graph<T>> graph, std::vector<std::string> const &input_node_names,
-    std::string const &label_node_name, std::string const &output_node_name,
-    fetch::ml::optimisers::LearningRateParam<DataType> const &learning_rate_param,
-    DataType const &                                          momentum_update)
-  : Optimiser<T>(graph, input_node_names, label_node_name, output_node_name, learning_rate_param)
-  , momentum_update_(momentum_update)
-{
-  Init();
-}
-
-// private
-
-template <class T>
-void MomentumOptimiser<T>::ApplyGradients(SizeType batch_size)
-{
-  auto trainable_it = this->graph_trainables_.begin();
-  auto gradient_it  = this->gradients_.begin();
-  auto mit          = momentum_.begin();
-
-  while (gradient_it != this->gradients_.end())
-  {
-    // Skip frozen trainables
-    if (!(*trainable_it)->GetFrozenState())
-    {
-
-      // momentum[i] = momentum_update * momentum[i] + learning_rate * (input_grad[i]/batch_size)
-      fetch::math::Multiply(*mit, momentum_update_, *mit);
-      fetch::math::Multiply((*trainable_it)->GetGradientsReferences(),
-                            (this->learning_rate_) / (static_cast<DataType>(batch_size)),
-                            *gradient_it);
-      fetch::math::Add(*mit, *gradient_it, *mit);
-
-      // output_grad[i]=-momentum[i]
-      fetch::math::Multiply(*mit, negative_one_, *gradient_it);
-
-      // we need to explicitly reset the gradients for this shared op to avoid double counting
-      // in the case of shared ops
-      (*trainable_it)->ResetGradients();
-    }
-
-    ++trainable_it;
-    ++gradient_it;
-    ++mit;
-  }
-
-  // calling apply gradients on the graph ensures that the node caches are reset properly
-  this->graph_->ApplyGradients(this->gradients_);
-}
-
-template <class T>
-void MomentumOptimiser<T>::ResetMomentum()
-{
-  for (auto &moment : this->momentum_)
-  {
-    moment.Fill(zero_);
-  }
-}
-
 }  // namespace optimisers
 }  // namespace ml
 }  // namespace fetch
